Perform update of the notebook controller without surge
This allows updating the notebook controllers without increasing the
resources requirements of ODH during the update, which means it can
better tolerates situation such as the cluster being full.

This means the notebook controller will be unavailable during the
update. However, this has not visible consequences except for a slight
delay in the notebooks CR reconciliation, until the new controller pod
start reconciling.